[Keyboard] Fixup Crkbd default keymap
