+++ keyboards/crkbd/info.json
@@ -1,21 +1,35 @@
 {
-  "keyboard_name": "Corne",
-  "manufacturer": "foostan",
-  "url": "",
-  "maintainer": "qmk",
-  "usb": {
-    "vid": "0x4653",
-    "pid": "0x0001",
-    "device_version": "0.0.1"
-  },
-  "rgb_matrix": {
-    "driver": "WS2812"
-  },
-  "matrix_pins": {
-    "cols": ["F4", "F5", "F6", "F7", "B1", "B3"],
-    "rows": ["D4", "C6", "D7", "E6"]
-  },
-  "diode_direction": "COL2ROW",
-  "processor": "atmega32u4",
-  "community_layouts": ["split_3x5_3", "split_3x6_3"]
+    "keyboard_name": "Corne",
+    "manufacturer": "foostan",
+    "url": "",
+    "maintainer": "qmk",
+    "usb": {
+        "vid": "0x4653",
+        "pid": "0x0001",
+        "device_version": "0.0.1"
+    },
+    "rgb_matrix": {
+        "driver": "WS2812"
+    },
+    "features": {
+        "bootmagic": true,
+        "extrakey": true,
+        "lto": true,
+        "mousekey": true,
+        "nkro": true,
+        "oled": true
+    },
+    "build": {
+        "lto": true
+    },
+    "matrix_pins": {
+        "cols": [ "F4", "F5", "F6", "F7", "B1", "B3" ],
+        "rows": [ "D4", "C6", "D7", "E6" ]
+    },
+    "diode_direction": "COL2ROW",
+    "split": {
+        "enabled": true,
+    },
+    "processor": "atmega32u4",
+    "community_layouts": [ "split_3x5_3", "split_3x6_3" ]
 }
