+++ keyboards/crkbd/keymaps/default/config.h
@@ -47,5 +47,3 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
     #define RGBLIGHT_SAT_STEP 17
     #define RGBLIGHT_VAL_STEP 17
 #endif
-
-#define OLED_FONT_H "keyboards/crkbd/lib/glcdfont.c"
