+++ keyboards/crkbd/keymaps/default/rules.mk
@@ -1,5 +1,1 @@
-MOUSEKEY_ENABLE = yes    # Mouse keys
 RGBLIGHT_ENABLE = yes    # Enable WS2812 RGB underlight.
-OLED_ENABLE     = yes
-OLED_DRIVER     = SSD1306
-LTO_ENABLE      = yes
