--- keyboards/crkbd/r2g/rules.mk
@@ -1,6 +1,0 @@
-OLED_ENABLE     = yes
-OLED_DRIVER     = SSD1306
-LTO_ENABLE      = yes
-RGBLIGHT_ENABLE = yes
-
-SPLIT_KEYBOARD = yes
