+++ keyboards/crkbd/rev1/info.json
@@ -1,11 +1,72 @@
 {
+    "features": {
+        "rgblight": true
+    },
     "split": {
         "soft_serial_pin": "D2"
     },
     "ws2812": {
         "pin": "D3"
     },
-    "bootloader": "caterina",
+    "rgb_matrix": {
+        "layout": [
+            {"x": 85, "y": 16, "flags": 2},
+            {"x": 50, "y": 13, "flags": 2},
+            {"x": 16, "y": 20, "flags": 2},
+            {"x": 16, "y": 38, "flags": 2},
+            {"x": 50, "y": 48, "flags": 2},
+            {"x": 85, "y": 52, "flags": 2},
+            {"matrix": [3, 5], "x": 95, "y": 63, "flags": 1},
+            {"matrix": [2, 5], "x": 85, "y": 39, "flags": 4},
+            {"matrix": [1, 5], "x": 85, "y": 21, "flags": 4},
+            {"matrix": [0, 5], "x": 85, "y": 4, "flags": 4},
+            {"matrix": [0, 4], "x": 68, "y": 2, "flags": 4},
+            {"matrix": [1, 4], "x": 68, "y": 19, "flags": 4},
+            {"matrix": [2, 4], "x": 68, "y": 37, "flags": 4},
+            {"matrix": [3, 4], "x": 80, "y": 58, "flags": 1},
+            {"matrix": [3, 3], "x": 60, "y": 55, "flags": 1},
+            {"matrix": [2, 3], "x": 50, "y": 35, "flags": 4},
+            {"matrix": [1, 3], "x": 50, "y": 13, "flags": 4},
+            {"matrix": [0, 3], "x": 50, "y": 0, "flags": 4},
+            {"matrix": [0, 2], "x": 33, "y": 3, "flags": 4},
+            {"matrix": [1, 2], "x": 33, "y": 20, "flags": 4},
+            {"matrix": [2, 2], "x": 33, "y": 37, "flags": 4},
+            {"matrix": [2, 1], "x": 16, "y": 42, "flags": 4},
+            {"matrix": [1, 1], "x": 16, "y": 24, "flags": 4},
+            {"matrix": [0, 1], "x": 16, "y": 7, "flags": 4},
+            {"matrix": [0, 0], "x": 0, "y": 7, "flags": 1},
+            {"matrix": [1, 0], "x": 0, "y": 24, "flags": 1},
+            {"matrix": [2, 0], "x": 0, "y": 41, "flags": 1},
+            {"x": 139, "y": 16, "flags": 2},
+            {"x": 174, "y": 13, "flags": 2},
+            {"x": 208, "y": 20, "flags": 2},
+            {"x": 208, "y": 38, "flags": 2},
+            {"x": 174, "y": 48, "flags": 2},
+            {"x": 139, "y": 52, "flags": 2},
+            {"matrix": [7, 5], "x": 129, "y": 63, "flags": 1},
+            {"matrix": [6, 5], "x": 139, "y": 39, "flags": 4},
+            {"matrix": [5, 5], "x": 139, "y": 21, "flags": 4},
+            {"matrix": [4, 5], "x": 139, "y": 4, "flags": 4},
+            {"matrix": [4, 4], "x": 156, "y": 2, "flags": 4},
+            {"matrix": [5, 4], "x": 156, "y": 19, "flags": 4},
+            {"matrix": [6, 4], "x": 156, "y": 37, "flags": 4},
+            {"matrix": [7, 4], "x": 144, "y": 58, "flags": 1},
+            {"matrix": [7, 3], "x": 164, "y": 55, "flags": 1},
+            {"matrix": [6, 3], "x": 174, "y": 35, "flags": 4},
+            {"matrix": [5, 3], "x": 174, "y": 13, "flags": 4},
+            {"matrix": [4, 3], "x": 174, "y": 0, "flags": 4},
+            {"matrix": [4, 2], "x": 191, "y": 3, "flags": 4},
+            {"matrix": [5, 2], "x": 191, "y": 20, "flags": 4},
+            {"matrix": [6, 2], "x": 191, "y": 37, "flags": 4},
+            {"matrix": [6, 1], "x": 208, "y": 42, "flags": 4},
+            {"matrix": [5, 1], "x": 208, "y": 24, "flags": 4},
+            {"matrix": [4, 1], "x": 208, "y": 7, "flags": 4},
+            {"matrix": [4, 0], "x": 224, "y": 7, "flags": 1},
+            {"matrix": [5, 0], "x": 224, "y": 24, "flags": 1},
+            {"matrix": [6, 0], "x": 224, "y": 41, "flags": 1}
+        ]
+    },
+    "development_board": "promicro",
     "layout_aliases": {
         "LAYOUT": "LAYOUT_split_3x6_3"
     },
