--- keyboards/crkbd/rev1/legacy/rules.mk
@@ -1,1 +1,0 @@
-DEFAULT_FOLDER = crkbd/rev1
