+++ keyboards/crkbd/rev1/rules.mk
@@ -1,1 +1,1 @@
-SPLIT_KEYBOARD = yes
+RGBLIGHT_ENABLE = yes
